requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.0.0
django-rest-swagger==2.1.1
django2-rest-passwordreset==1.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==3.3.1
